+++ mypy/metastore.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     import sqlite3
 
 
-class MetadataStore:
+class MetadataStore:  # type: ignore[abstract]
     """Generic interface for metadata storage."""
 
     @abstractmethod