+++ mypy/nodes.py
@@ -247,7 +247,7 @@ Lvalue: _TypeAlias = Expression
 
 
 @trait
-class SymbolNode(Node):
+class SymbolNode(Node):  # type: ignore[abstract]
     """Nodes that can be stored in a symbol table."""
 
     __slots__ = ()
@@ -505,7 +505,7 @@ class ImportAll(ImportBase):
 FUNCBASE_FLAGS: Final = ["is_property", "is_class", "is_static", "is_final", "is_type_check_only"]
 
 
-class FuncBase(Node):
+class FuncBase(Node):  # type: ignore[abstract]
     """Abstract base class for function-like nodes.
 
     N.B: Although this has SymbolNode subclasses (FuncDef,
@@ -710,7 +710,7 @@ FUNCITEM_FLAGS: Final = FUNCBASE_FLAGS + [
 ]
 
 
-class FuncItem(FuncBase):
+class FuncItem(FuncBase):  # type: ignore[abstract]
     """Base class for nodes usable as overloaded function items."""
 
     __slots__ = (
@@ -1021,6 +1021,7 @@ class Var(SymbolNode):
         "is_settable_property",
         "is_classvar",
         "is_abstract_var",
+        "is_read_only",
         "is_final",
         "is_index_var",
         "final_unset_in_class",
@@ -1058,6 +1059,7 @@ class Var(SymbolNode):
         self.is_classvar = False
         self.is_abstract_var = False
         self.is_index_var = False
+        self.is_read_only = False
         # Set to true when this variable refers to a module we were unable to
         # parse for some reason (eg a silenced module)
         self.is_suppressed_import = False
@@ -2558,7 +2560,7 @@ CONTRAVARIANT: Final = 2
 VARIANCE_NOT_READY: Final = 3  # Variance hasn't been inferred (using Python 3.12 syntax)
 
 
-class TypeVarLikeExpr(SymbolNode, Expression):
+class TypeVarLikeExpr(SymbolNode, Expression):  # type: ignore[abstract]
     """Base class for TypeVarExpr, ParamSpecExpr and TypeVarTupleExpr.
 
     Note that they are constructed by the semantic analyzer.