+++ mypy/plugin.py
@@ -151,7 +151,7 @@ from mypy.types import (
 
 
 @trait
-class TypeAnalyzerPluginInterface:
+class TypeAnalyzerPluginInterface:  # type: ignore[abstract]
     """Interface for accessing semantic analyzer functionality in plugins.
 
     Methods docstrings contain only basic info. Look for corresponding implementation
@@ -195,7 +195,7 @@ class AnalyzeTypeContext(NamedTuple):
 
 
 @mypyc_attr(allow_interpreted_subclasses=True)
-class CommonPluginApi:
+class CommonPluginApi:  # type: ignore[abstract]
     """
     A common plugin API (shared between semantic analysis and type checking phases)
     that all plugin hooks get independently of the context.
@@ -217,7 +217,7 @@ class CommonPluginApi:
 
 
 @trait
-class CheckerPluginInterface:
+class CheckerPluginInterface:  # type: ignore[abstract]
     """Interface for accessing type checker functionality in plugins.
 
     Methods docstrings contain only basic info. Look for corresponding implementation
@@ -254,7 +254,7 @@ class CheckerPluginInterface:
 
 
 @trait
-class SemanticAnalyzerPluginInterface:
+class SemanticAnalyzerPluginInterface:  # type: ignore[abstract]
     """Interface for accessing semantic analyzer functionality in plugins.
 
     Methods docstrings contain only basic info. Look for corresponding implementation