+++ mypyc/analysis/dataflow.py
@@ -169,7 +169,7 @@ class AnalysisResult(Generic[T]):
 GenAndKill = Tuple[Set[T], Set[T]]
 
 
-class BaseAnalysisVisitor(OpVisitor[GenAndKill[T]]):
+class BaseAnalysisVisitor(OpVisitor[GenAndKill[T]]):  # type: ignore[abstract]
     def visit_goto(self, op: Goto) -> GenAndKill[T]:
         return set(), set()